+++ survival-client/src/main/java/net/survival/client/graphics/blockrenderer/DefaultBlockRenderer.java
@@ -9,7 +9,7 @@ class DefaultBlockRenderer extends BlockRenderer
     }
 
     @Override
-    public void pushTopFaces(int x, int y, int z, short blockID, short adjacentBlockID,
+    public void pushTopFaces(int x, int y, int z, int blockID, int adjacentBlockID,
             GLDisplayList.Builder builder)
     {
         if (blockToBlockingBottomTable[adjacentBlockID])
@@ -28,7 +28,7 @@ class DefaultBlockRenderer extends BlockRenderer
     }
 
     @Override
-    public void pushBottomFaces(int x, int y, int z, short blockID, short adjacentBlockID,
+    public void pushBottomFaces(int x, int y, int z, int blockID, int adjacentBlockID,
             GLDisplayList.Builder builder)
     {
         if (blockToBlockingTopTable[adjacentBlockID])
@@ -47,7 +47,7 @@ class DefaultBlockRenderer extends BlockRenderer
     }
 
     @Override
-    public void pushLeftFaces(int x, int y, int z, short blockID, short adjacentBlockID,
+    public void pushLeftFaces(int x, int y, int z, int blockID, int adjacentBlockID,
             GLDisplayList.Builder builder)
     {
         if (blockToBlockingRightTable[adjacentBlockID])
@@ -66,7 +66,7 @@ class DefaultBlockRenderer extends BlockRenderer
     }
 
     @Override
-    public void pushRightFaces(int x, int y, int z, short blockID, short adjacentBlockID,
+    public void pushRightFaces(int x, int y, int z, int blockID, int adjacentBlockID,
             GLDisplayList.Builder builder)
     {
         if (blockToBlockingLeftTable[adjacentBlockID])
@@ -85,7 +85,7 @@ class DefaultBlockRenderer extends BlockRenderer
     }
 
     @Override
-    public void pushFrontFaces(int x, int y, int z, short blockID, short adjacentBlockID,
+    public void pushFrontFaces(int x, int y, int z, int blockID, int adjacentBlockID,
             GLDisplayList.Builder builder)
     {
         if (blockToBlockingBackTable[adjacentBlockID])
@@ -104,8 +104,8 @@ class DefaultBlockRenderer extends BlockRenderer
     }
 
     @Override
-    public void pushBackFaces(int x, int y, int z, short blockID,
-            short adjacentBlockID, GLDisplayList.Builder builder)
+    public void pushBackFaces(int x, int y, int z, int blockID, int adjacentBlockID,
+            GLDisplayList.Builder builder)
     {
         if (blockToBlockingFrontTable[adjacentBlockID])
             return;
